feat: full migrate to ts ready
