+++ package.json
@@ -15,19 +15,15 @@
     },
     "author": "Sergey Grabak",
     "license": "ISC",
-    "jest": {
-        "verbose": true,
-        "testEnvironmentOptions": {
-            "url": "http://localhost/"
-        },
-        "testEnvironment": "jsdom"
-    },
     "main": "src/init.js",
     "devDependencies": {
         "@babel/core": "^7.20.12",
         "@babel/eslint-parser": "^7.19.1",
         "@babel/preset-env": "^7.20.2",
+        "@babel/preset-typescript": "^7.21.5",
+        "@jest/globals": "^29.5.0",
         "@testing-library/dom": "^9.2.0",
+        "@types/jest": "^29.5.2",
         "babel-jest": "^29.3.1",
         "debug": "^4.3.4",
         "eslint": "^8.39.0",
@@ -50,6 +46,7 @@
         "jest-environment-jsdom": "^29.3.1",
         "lint-staged": "^13.2.2",
         "prettier": "^2.8.8",
+        "ts-jest": "^29.1.0",
         "typescript": "^5.0.2",
         "vite": "^4.3.9"
     },
