+++ src/components/Key.ts
@@ -1,13 +1,17 @@
-const Key = (symbols, code, types = null) => {
+const Key = (
+    symbols: string[],
+    code: string,
+    types: string[] | null = null
+) => {
     const keyContainer = document.createElement('span');
     keyContainer.dataset.code = code;
 
     if (types && types.includes('multi')) {
-        keyContainer.dataset.multi = true;
+        keyContainer.dataset.multi = 'true';
     }
 
     if (code.includes('Key') || (types && types.includes('single'))) {
-        keyContainer.dataset.single = true;
+        keyContainer.dataset.single = 'true';
     }
 
     keyContainer.classList.add('key');
@@ -17,8 +21,9 @@ const Key = (symbols, code, types = null) => {
         keyContainer.classList.add(...types);
     }
 
-    const mainSymbol = symbols[0].toLowerCase();
-    const secondarySymbol = symbols.length > 1 ? symbols[1] : undefined;
+    const mainSymbol: string = symbols[0].toLowerCase();
+    const secondarySymbol: string | undefined =
+        symbols.length > 1 ? symbols[1] : undefined;
 
     keyContainer.innerHTML =
         symbols.length > 1
