+++ src/helpers/consts.ts
@@ -1,0 +1,20 @@
+export const Actions: Record<string, string> = {
+    Insert: 'Insert',
+    Backspace: 'Backspace',
+    Delete: 'Delete',
+    Tab: 'Tab',
+    Enter: 'Enter',
+    ShiftRight: 'ShiftRight',
+    ShiftLeft: 'ShiftLeft',
+    Space: 'Space',
+    AltRight: 'AltRight',
+    AltLeft: 'AltLeft',
+    MetaLeft: 'MetaLeft',
+    CapsLock: 'CapsLock',
+    ControlRight: 'ControlRight',
+    ControlLeft: 'ControlLeft',
+    ArrowLeft: 'ArrowLeft',
+    ArrowRight: 'ArrowRight',
+    ArrowUp: 'ArrowUp',
+    ArrowDown: 'ArrowDown',
+};
