+++ vite.config.js
@@ -5,6 +5,7 @@ export default defineConfig({
     resolve: {
         alias: {
             '@components': path.resolve(__dirname, './src/components'),
+            '@helpers': path.resolve(__dirname, './src/helpers'),
         },
     },
 });
